requirements.txt
flask==3.0.*
flask-login==0.6.*
flask-sqlalchemy==3.1.*
flask-mail==0.10.*
gunicorn==21.*
paramiko==3.4.*
requests==2.31.*
psutil==5.9.*
python-dotenv==1.0.*
